VECTRA/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pinecone
openai
python-dotenv
requests
pydantic
sentence-transformers
